requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
cycler==0.10.0            # via matplotlib
kiwisolver==1.1.0         # via matplotlib
matplotlib==3.1.1
numpy==1.17.3
pyparsing==2.4.4          # via matplotlib
pyside2==5.13.2
python-dateutil==2.8.1    # via matplotlib
pyzmq==18.1.0             # via zmq
ruamel.yaml.clib==0.2.0   # via ruamel.yaml
ruamel.yaml==0.16.5
shiboken2==5.13.2         # via pyside2
six==1.13.0               # via cycler, python-dateutil
zmq==0.0.0

# The following packages are considered to be unsafe in a requirements file:
# setuptools==42.0.1        # via kiwisolver
